added add new file to collection actions
